made code even cleaner and reomved typos

+++ src/error.cpp
@@ -10,7 +10,7 @@ void error::syntaxError(const std::string& error) // Print out syntax error
 	while(errorLine.at(0) == ' ' || errorLine.at(0) == '	')
 		errorLine.erase(errorLine.begin());
 	std::cerr << errorLine << std::endl; // find line of error
-	for(unsigned int i = 0; i < lexer::tokens.at(parser::tokCount).pos; i++) // point to error token
+	for(unsigned long i = 0; i < lexer::tokens.at(parser::tokCount).pos; i++) // point to error token
 		std::cerr << " ";
 	std::cerr << "^" << "\033[0m" << std::endl; // reset to white
